[filter] Moved Filter file to pycram src
