+++ src/pycram/testing.py
@@ -10,7 +10,7 @@ from .robot_description import RobotDescription, RobotDescriptionManager
 from .process_module import ProcessModule
 from .datastructures.enums import ObjectType, WorldMode
 from .object_descriptors.urdf import ObjectDescription
-from .ros import  VizMarkerPublisher
+from .ros_utils.viz_marker_publisher import VizMarkerPublisher
 from pycrap import ontology, Milk, Robot, Kitchen, Cereal
 import owlready2
